refactor(admin-web): #521 — split 867 LoC SystemAnnouncementsPage into focused components, add validator + tests
