+++ frontend/apps/ppt-web/src/App.tsx
@@ -19,7 +19,6 @@ import {
   useAnnouncementAcknowledgmentStats,
   useAnnouncementComments,
   useAnnouncements,
-  usePinnedAnnouncements,
   useArchiveAnnouncement,
   useBuildings,
   useCancelOutage,
@@ -39,6 +38,7 @@ import {
   useOutages,
   usePauseSchedule,
   usePinAnnouncement,
+  usePinnedAnnouncements,
   usePublishAnnouncement,
   useReportExecutionHistory,
   useResolveOutage,
